Remove unnecessary quickpulse interfaces
