+++ core/src/main/java/com/microsoft/applicationinsights/internal/quickpulse/QuickPulseCoordinator.java
@@ -21,4 +21,123 @@
 
 package com.microsoft.applicationinsights.internal.quickpulse;
 
-interface QuickPulseCoordinator {}
+import com.microsoft.applicationinsights.internal.util.LocalStringsUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class QuickPulseCoordinator implements Runnable {
+
+  private static final Logger logger = LoggerFactory.getLogger(QuickPulseCoordinator.class);
+  private String qpsServiceRedirectedEndpoint;
+  private long qpsServicePollingIntervalHintMillis;
+
+  private volatile boolean stopped = false;
+  private volatile boolean pingMode = true;
+
+  private final QuickPulsePingSender pingSender;
+  private final QuickPulseDataFetcher dataFetcher;
+  private final QuickPulseDataSender dataSender;
+
+  private final long waitBetweenPingsInMillis;
+  private final long waitBetweenPostsInMillis;
+  private final long waitOnErrorInMillis;
+
+  public QuickPulseCoordinator(QuickPulseCoordinatorInitData initData) {
+    dataSender = initData.dataSender;
+    pingSender = initData.pingSender;
+    dataFetcher = initData.dataFetcher;
+
+    waitBetweenPingsInMillis = initData.waitBetweenPingsInMillis;
+    waitBetweenPostsInMillis = initData.waitBetweenPostsInMillis;
+    waitOnErrorInMillis = initData.waitBetweenPingsInMillis;
+    qpsServiceRedirectedEndpoint = null;
+    qpsServicePollingIntervalHintMillis = -1;
+  }
+
+  @Override
+  public void run() {
+    try {
+      while (!stopped) {
+        long sleepInMillis;
+        if (pingMode) {
+          sleepInMillis = ping();
+        } else {
+          sleepInMillis = sendData();
+        }
+        Thread.sleep(sleepInMillis);
+      }
+    } catch (InterruptedException e) {
+      Thread.currentThread().interrupt();
+    } catch (ThreadDeath td) {
+      throw td;
+    } catch (Throwable t) {
+      // chomp
+    }
+  }
+
+  private long sendData() {
+    dataFetcher.prepareQuickPulseDataForSend(qpsServiceRedirectedEndpoint);
+    QuickPulseHeaderInfo currentQuickPulseHeaderInfo = dataSender.getQuickPulseHeaderInfo();
+
+    this.handleReceivedHeaders(currentQuickPulseHeaderInfo);
+
+    switch (currentQuickPulseHeaderInfo.getQuickPulseStatus()) {
+      case ERROR:
+        pingMode = true;
+        return waitOnErrorInMillis;
+
+      case QP_IS_OFF:
+        pingMode = true;
+        return qpsServicePollingIntervalHintMillis > 0
+            ? qpsServicePollingIntervalHintMillis
+            : waitBetweenPingsInMillis;
+
+      case QP_IS_ON:
+        return waitBetweenPostsInMillis;
+    }
+
+    logger.error("Critical error while sending QP data: unknown status, aborting");
+    QuickPulseDataCollector.INSTANCE.disable();
+    stopped = true;
+    return 0;
+  }
+
+  private long ping() {
+    QuickPulseHeaderInfo pingResult = pingSender.ping(qpsServiceRedirectedEndpoint);
+    this.handleReceivedHeaders(pingResult);
+    switch (pingResult.getQuickPulseStatus()) {
+      case ERROR:
+        return waitOnErrorInMillis;
+
+      case QP_IS_ON:
+        pingMode = false;
+        dataSender.startSending();
+        return waitBetweenPostsInMillis;
+      case QP_IS_OFF:
+        return qpsServicePollingIntervalHintMillis > 0
+            ? qpsServicePollingIntervalHintMillis
+            : waitBetweenPingsInMillis;
+    }
+
+    logger.error("Critical error while ping QP: unknown status, aborting");
+    QuickPulseDataCollector.INSTANCE.disable();
+    stopped = true;
+    return 0;
+  }
+
+  private void handleReceivedHeaders(QuickPulseHeaderInfo currentQuickPulseHeaderInfo) {
+    String redirectLink = currentQuickPulseHeaderInfo.getQpsServiceEndpointRedirect();
+    if (!LocalStringsUtils.isNullOrEmpty(redirectLink)) {
+      qpsServiceRedirectedEndpoint = redirectLink;
+    }
+
+    long newPollingInterval = currentQuickPulseHeaderInfo.getQpsServicePollingInterval();
+    if (newPollingInterval > 0) {
+      qpsServicePollingIntervalHintMillis = newPollingInterval;
+    }
+  }
+
+  public void stop() {
+    stopped = true;
+  }
+}
